refactor(user_service): rename infraestructure → infrastructure package
