+++ src/main/java/com/renewsim/backend/user_service/application/service/GetUserService.java
@@ -11,7 +11,7 @@ import com.renewsim.backend.user_service.application.port.in.GetUserUseCase;
 import com.renewsim.backend.user_service.application.port.out.UserRepositoryPort;
 import com.renewsim.backend.user_service.domain.model.User;
 import com.renewsim.backend.user_service.dto.UserResponse;
-import com.renewsim.backend.user_service.infraestructure.mapper.UserServiceMapper;
+import com.renewsim.backend.user_service.infrastructure.mapper.UserServiceMapper;
 
 @Slf4j
 @Service
