+++ src/main/java/com/renewsim/backend/user_service/infrastructure/client/RoleServiceClient.java
@@ -1,4 +1,4 @@
-package com.renewsim.backend.user_service.infraestructure.client;
+package com.renewsim.backend.user_service.infrastructure.client;
 
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
