+++ src/main/java/com/renewsim/backend/user_service/infrastructure/mapper/UserServiceMapper.java
@@ -1,9 +1,9 @@
-package com.renewsim.backend.user_service.infraestructure.mapper;
+package com.renewsim.backend.user_service.infrastructure.mapper;
 
 import com.renewsim.backend.user_service.domain.model.User;
 import com.renewsim.backend.user_service.dto.UserCreateRequest;
 import com.renewsim.backend.user_service.dto.UserResponse;
-import com.renewsim.backend.user_service.infraestructure.persistence.entity.UserEntity;
+import com.renewsim.backend.user_service.infrastructure.persistence.entity.UserEntity;
 import com.renewsim.backend.shared.domain.vo.RoleName;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
